+++ CHANGELOG.md
@@ -3,9 +3,9 @@ A more detailed list of changes is available in the corresponding milestones for
 
 
 ## 0.8.0 (2020-Jul-??)
-
 ### Bugfixes
-  - **com.google.fonts/check/description/broken_links**: Skip when html does not parse. (issue #2664)
+  - **[com.google.fonts/check/description/broken_links]**: Skip when html does not parse. (issue #2664)
+  - Checks if GSUB lookup format is 1 for ligature collection in `profiles/shared_conditions.py`; format 1 is the only significant one for `ligatures()` and `ligature_glyphs()`)
 
 
 ## 0.7.28 (2020-Jul-09)